refactory: move things around a bit
